account_service/accounts/views.py
from rest_framework import generics
from accounts.models import User
from accounts.serializers import UserSerializer


class UserList(generics.ListCreateAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer
